feat: add debug mode fetch handler
it's not currently possible to see the logs of the tail worker itself using the cloudflare dashboard or `wrangler tail`.

this PR adds a debug fetch handler so you can check that loki is satisfied with your offerings

License: MIT
