+++ example.dev.vars
@@ -1,0 +1,5 @@
+## full url of of the push enpoint on you loki instance
+LOKI_URL="https://xyz.grafana.net/loki/api/v1/push"
+
+## base64 encoded basic auth token
+LOKI_TOKEN=""
